fixing a wrong parser error

+++ SheetHandler/Hole.java
@@ -15,7 +15,7 @@ import Parser.GCode.NGCDocument;
  * Holes are essentially identical to Parts besides drawing
  */
 public class Hole extends Part {
-    public static final double HEAD_SIZE = Integer.parseInt(Settings.settings.get("ScrewHeadSize"));
+    public static final double HEAD_SIZE = Double.parseDouble(Settings.settings.get("ScrewHeadSize"));
 
     public Hole(File holeFile, double x, double y, double rot) {
         super(holeFile, x, y, rot);
